move buttons and stats to above tabs
