+++ nerfstudio/viewer_beta/render_state_machine.py
@@ -246,6 +246,8 @@ class RenderStateMachine(threading.Thread):
             jpeg_quality=self.viewer.config.jpeg_quality,
             depth=depth,
         )
+        res = f"{selected_output.shape[0]}x{selected_output.shape[1]}"
+        self.viewer.stats_markdown.content = self.viewer.make_stats_markdown(None,res)
 
     def _calculate_image_res(self, aspect_ratio: float) -> Tuple[int, int]:
         """Calculate the maximum image height that can be rendered in the time budget
